# docs/swagger.yaml
definitions:
  github_com_maczh_mgin_models.Result-any:
    properties:
      data: {}
      msg:
        type: string
      page:
        $ref: '#/definitions/models.ResultPage'
      status:
        type: integer
    type: object
...
  github_com_maczh_mgin_models.Result-github_com_maczh_mgin_models_sys_SysDict:
    properties:
      data:
        $ref: '#/definitions/github_com_maczh_mgin_models_sys.SysDict'
      msg:
        type: string
      page:
        $ref: '#/definitions/models.ResultPage'
      status:
        type: integer
    type: object
  github_com_maczh_mgin_models.Result-github_com_maczh_mgin_models_sys_SysMenu:
    properties:
      data:
        $ref: '#/definitions/github_com_maczh_mgin_models_sys.SysMenu'
      msg:
        type: string
      page:
        $ref: '#/definitions/models.ResultPage'
      status:
        type: integer
    type: object
...
  github_com_maczh_mgin_models.ResultPage:
    properties:
      count:
        description: 总页数
        type: integer
      index:
        description: 页号
        type: integer
      size:
        description: 分页大小
        type: integer
      total:
        description: 总记录数
        type: integer
    type: object
  github_com_maczh_mgin_models_sys.SysApi:
    properties:
      api_group:
        type: string
      api_path:
        type: string
      createAt:
        type: string
      createBy:
        type: string
      delFlag:
        type: integer
      description:
        type: string
      enabled:
        type: integer
      id:
        type: integer
      method:
        type: string
      name:
        type: string
      need_auth:
        type: integer
      need_log:
        type: integer
      tenantId:
        type: integer
      updateAt:
        type: string
      updateBy:
        type: string
    type: object
...
  github_com_maczh_mgin_models_sys.SysDict:
    properties:
      createAt:
        type: string
      createBy:
        type: string
      delFlag:
        type: integer
      id:
        description: 主键ID
        type: integer
      key:
        description: 字典键名
        type: string
      name:
        description: 字典名称
        type: string
      parent_id:
        description: 父级ID
        type: integer
      remark:
        description: 备注
        type: string
      sort:
        description: 排序
        type: integer
      tenantId:
        type: integer
      type:
        description: 字典类型
        type: string
      updateAt:
        type: string
      updateBy:
        type: string
      value:
        description: 字典值
        type: string
    type: object
  github_com_maczh_mgin_models_sys.SysMenu:
    properties:
      active_menu:
        type: string
      affix:
        type: integer
      always_show:
        type: integer
      breadcrumb:
        type: integer
      children;omitempty:
        description: 关联关系
        items:
          $ref: '#/definitions/github_com_maczh_mgin_models_sys.SysMenu'
        type: array
      component:
        type: string
      createAt:
        type: string
      createBy:
        type: string
      delFlag:
        type: integer
      hidden:
        type: integer
      icon:
        type: string
      id:
        type: integer
      keep_alive:
        type: integer
      name:
        type: string
      parent;omitempty:
        allOf:
        - $ref: '#/definitions/github_com_maczh_mgin_models_sys.SysMenu'
        description: 自关联
      parent_id:
        type: integer
      path:
        type: string
      redirect:
        type: string
      roles;omitempty:
        description: 多对多关联
        items:
          $ref: '#/definitions/github_com_maczh_mgin_models_sys.SysRole'
        type: array
      sort:
        type: integer
      status:
        type: integer
      tenantId:
        type: integer
      title:
        type: string
      updateAt:
        type: string
      updateBy:
        type: string
    type: object
...
  github_com_maczh_mgin_models_sys.SysRole:
    properties:
      apis;omitempty:
        items:
          $ref: '#/definitions/github_com_maczh_mgin_models_sys.SysApi'
        type: array
      createAt:
        type: string
      createBy:
        type: string
      delFlag:
        type: integer
      description:
        type: string
      id:
        type: integer
      isEnable:
        type: integer
      menus;omitempty:
        items:
          $ref: '#/definitions/github_com_maczh_mgin_models_sys.SysMenu'
        type: array
      roleIdent:
        type: string
      roleName:
        type: string
      tenantId:
        type: integer
      updateAt:
        type: string
      updateBy:
        type: string
      users;omitempty:
        description: 关联关系
        items:
          $ref: '#/definitions/github_com_maczh_mgin_models_sys.SysUser'
        type: array
    type: object
  github_com_maczh_mgin_models_sys.SysUser:
    properties:
      avatar:
        type: string
      createAt:
        type: string
      createBy:
        type: string
      delFlag:
        type: integer
      email:
        type: string
      id:
        type: integer
      loginDate:
        type: string
      loginIp:
        type: string
      loginName:
        type: string
      mobile:
        type: string
      nickName:
        type: string
      password:
        type: string
      remark:
        type: string
      sex:
        type: integer
      status:
        type: integer
      tenantId:
        type: integer
      updateAt:
        type: string
      updateBy:
        type: string
      userType:
        type: string
    type: object
...
  github_com_maczh_mgin_models_sys_request.CreateDictReq:
    properties:
      key:
        type: string
      name:
        type: string
      parentId:
        type: integer
      remark:
        type: string
      sort:
        type: integer
      type:
        type: string
      value:
        type: string
    required:
    - key
    - name
    - type
    - value
    type: object
  github_com_maczh_mgin_models_sys_request.CreateMenuReq:
    properties:
      activeMenu:
        type: string
      affix:
        type: integer
      alwaysShow:
        type: integer
      breadcrumb:
        type: integer
      component:
        type: string
      hidden:
        type: integer
      icon:
        type: string
      keepAlive:
        type: integer
      name:
        type: string
      parentId:
        type: integer
      path:
        type: string
      redirect:
        type: string
      sort:
        type: integer
      status:
        type: integer
      title:
        type: string
    required:
    - component
    - name
    - path
    - title
    type: object
...
  github_com_maczh_mgin_models_sys_request.DeleteByIdReq:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  github_com_maczh_mgin_models_sys_request.GetDictReq:
    properties:
      id:
        type: integer
      key:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
...
  models.ResultPage:
    properties:
      count:
        description: 总页数
        type: integer
      index:
        description: 页号
        type: integer
      size:
        description: 分页大小
        type: integer
      total:
        description: 总记录数
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /api/v1/sys/dict/add:
    post:
      consumes:
      - application/json
      description: 新增一个系统字典信息
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 新增字典请求参数
        in: body
        name: dict
        required: true
        schema:
          $ref: '#/definitions/github_com_maczh_mgin_models_sys_request.CreateDictReq'
      produces:
      - application/json
      responses:
        "200":
          description: 新增成功，返回新增的字典信息
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-github_com_maczh_mgin_models_sys_SysDict'
        "500":
          description: 参数绑定失败或新增字典失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 新增字典
      tags:
      - 系统字典
  /api/v1/sys/dict/del:
    post:
      consumes:
      - application/json
      description: 根据 ID 删除系统字典信息
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 删除字典请求参数，包含字典 ID
        in: body
        name: dict
        required: true
        schema:
          $ref: '#/definitions/github_com_maczh_mgin_models_sys_request.DeleteByIdReq'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
        "500":
          description: 参数绑定失败或删除字典失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 删除字典
      tags:
      - 系统字典
  /api/v1/sys/dict/get:
    post:
      consumes:
      - application/json
      description: 根据 ID 或其他组合条件获取单个系统字典信息
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 获取单个字典请求参数
        in: body
        name: dict
        required: true
        schema:
          $ref: '#/definitions/github_com_maczh_mgin_models_sys_request.GetDictReq'
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功，返回单个字典信息
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-github_com_maczh_mgin_models_sys_SysDict'
        "400":
          description: 参数不完整
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
        "500":
          description: 参数绑定失败、参数不完整或获取字典失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 获取单个字典
      tags:
      - 系统字典
  /api/v1/sys/dict/list:
    get:
      consumes:
      - application/json
      description: 根据查询条件获取系统字典列表
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: name
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        name: pageSize
        type: integer
      - in: query
        name: parentId
        type: integer
      - in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功，返回字典列表和分页信息
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.ResultPage'
        "500":
          description: 参数绑定失败或获取字典列表失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 获取字典列表
      tags:
      - 系统字典
  /api/v1/sys/dict/update:
    post:
      consumes:
      - application/json
      description: 更新系统字典信息
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 更新后的字典信息
        in: body
        name: dict
        required: true
        schema:
          $ref: '#/definitions/github_com_maczh_mgin_models_sys.SysDict'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
        "500":
          description: 参数绑定失败或更新字典失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 更新字典
      tags:
      - 系统字典
  /api/v1/sys/menu/add:
    post:
      consumes:
      - application/json
      description: 新增一个系统菜单信息
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 新增菜单请求体
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/github_com_maczh_mgin_models_sys_request.CreateMenuReq'
      produces:
      - application/json
      responses:
        "200":
          description: 新增菜单成功，返回新增的菜单信息
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-github_com_maczh_mgin_models_sys_SysMenu'
        "500":
          description: 参数绑定失败或新增菜单失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 新增菜单
      tags:
      - 系统菜单
  /api/v1/sys/menu/del:
    post:
      consumes:
      - application/json
      description: 根据菜单 ID 删除对应的菜单信息
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除菜单成功
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
        "500":
          description: 参数绑定失败或删除菜单失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 删除菜单
      tags:
      - 系统菜单
  /api/v1/sys/menu/get:
    get:
      consumes:
      - application/json
      description: 根据请求参数获取单个菜单的详细信息
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: id
        type: integer
      - in: query
        name: path
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取菜单信息成功，返回菜单详细信息
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-github_com_maczh_mgin_models_sys_SysMenu'
        "500":
          description: 参数绑定失败或获取菜单信息失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 获取菜单信息
      tags:
      - 系统菜单
  /api/v1/sys/menu/list:
    get:
      consumes:
      - application/json
      description: 根据请求参数获取菜单列表，支持分页
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        name: pageSize
        type: integer
      - in: query
        name: parentId
        type: integer
      - in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取菜单列表成功，返回菜单列表和分页信息
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.ResultPage'
        "500":
          description: 参数绑定失败或获取菜单列表失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 获取菜单列表
      tags:
      - 系统菜单
  /api/v1/sys/menu/tree:
    get:
      consumes:
      - application/json
      description: 根据父菜单 ID 获取菜单的树形结构信息
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: parentId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取菜单树成功，返回菜单树形结构信息
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
        "500":
          description: 参数绑定失败或获取菜单树失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 获取菜单树
      tags:
      - 系统菜单
  /api/v1/sys/menu/update:
    post:
      consumes:
      - application/json
      description: 更新已存在的菜单信息
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 更新后的菜单信息
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/github_com_maczh_mgin_models_sys.SysMenu'
      produces:
      - application/json
      responses:
        "200":
          description: 更新菜单信息成功
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
        "500":
          description: 参数绑定失败或更新菜单信息失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 更新菜单信息
      tags:
      - 系统菜单
...
swagger: "2.0"